backend/requirements.txt
opencv-python==4.8.1.78
python-dotenv==1.0.0
easyocr==1.7.1
fastapi==0.104.1
motor==2.5.1
pymongo==3.12.0
python-multipart==0.0.6
uvicorn==0.24.0
numpy==1.24.3